+++ .gitignore
@@ -35,3 +35,5 @@ out/
 
 ### VS Code ###
 .vscode/
+.envrc
+*.sqlite.db
