+++ settings.gradle
@@ -1,32 +1,34 @@
 pluginManagement {
     repositories {
-        maven {
-            url 'https://maven.pkg.github.com/neverendingsupport/spring-boot'
-            credentials {
-                username = System.getenv("GITHUB_USER")
-                password = System.getenv("GITHUB_TOKEN")
-            }
-        }
+//        maven {
+//            url 'https://maven.pkg.github.com/neverendingsupport/spring-boot'
+//            credentials {
+//                username = System.getenv("GITHUB_USER")
+//                password = System.getenv("GITHUB_TOKEN")
+//            }
+//        }
+        mavenLocal()
         gradlePluginPortal()
         mavenCentral()
     }
-    resolutionStrategy {
-        eachPlugin {
-            println "****--------------------------------------****"
-            println "Requested Plugin Namespace: ${requested.id.namespace}"
-            println "Requested Plugin ID: ${requested.id.id}"
-            println "Requested Plugin Version: ${requested.version}"
-
-            if (requested.id.namespace == 'com.herodevs.nes.springframework') {
-                println "NAMESPACE MATCH 1"
-                useModule("com.herodevs.nes.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
-            }
-//            if (requested.id.id == 'com.herodevs.nes.springframework.boot') {
-//                println "NAMESPACE MATCH 2"
-//                useModule("com.herodevs.nes.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
-//            }
-        }
-    }
+//    resolutionStrategy {
+//        eachPlugin {
+////            println "****--------------------------------------****"
+////            println "Requested Plugin Namespace: ${requested.id.namespace}"
+////            println "Requested Plugin ID: ${requested.id.id}"
+////            println "Requested Plugin Version: ${requested.version}"
+//
+////            if (requested.id.namespace == 'com.herodevs.nes.springframework') {
+////                println "NAMESPACE MATCH 1 "
+////                useModule("com.herodevs.nes.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
+////                useModule("org.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
+////            }
+////            if (requested.id.id == 'com.herodevs.nes.springframework.boot') {
+////                println "NAMESPACE MATCH 2"
+////                useModule("com.herodevs.nes.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
+////            }
+//        }
+//    }
 }
