+++ src/main/resources/application.properties
@@ -1,7 +1,23 @@
-spring.jpa.open-in-view=false
-spring.jpa.hibernate.ddl-auto=none
-spring.datasource.url=jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;MODE=PostgreSQL;DATABASE_TO_LOWER=TRUE
-
+#debug=true
 logging.level.sql=DEBUG
-
 security.allowedOrigins=http://localhost:3000,http://localhost:3001
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect = com.enigmabridge.hibernate.dialect.SQLiteDialect
+spring.datasource.url=jdbc:sqlite:local.sqlite.db
+spring.datasource.driver-class-name=org.sqlite.JDBC
+spring.datasource.username=
+spring.datasource.password=
+spring.jackson.deserialization.UNWRAP_ROOT_VALUE=true
+
+spring.sql.init.platform=sqlite
+spring.sql.init.schema-locations=classpath:schema.sql
+spring.sql.init.data-locations=classpath:data.sql
+spring.datasource.initialization-mode=always
+
+# Location of schema SQL files
+#spring.datasource.schema=classpath:schema.sql
+
+# Location of data SQL files
+#spring.datasource.data=classpath:data.sql
+
